app/requirements.txt
pyyaml
requests
django-cors-headers
python-decouple
asgiref==3.8.1
Django==4.2.16
django-ckeditor
django-js-asset==2.2.0
djangorestframework==3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
django-filter==24.3
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
pyTelegramBotAPI==4.24.0
django-environ==0.11.2